Actualise interfaces

+++ core/window/WindowModule.hpp
@@ -36,9 +36,9 @@ public:
 	void closeWindow() override;
 	bool isOpen() override;
 	void clear() override;
-	void drawSprite(std::string asset, int color, std::pair<int, int> position) override;
-	void drawRectangle(int color, std::pair<int, int> position) override;
-	void drawText(std::string text, int color, std::pair<int, int> position) override;
+	void drawSprite(std::string asset, int color, std::pair<size_t, size_t> position) override;
+	void drawRectangle(int color, std::pair<size_t, size_t> position) override;
+	void drawText(std::string text, int color, std::pair<size_t, size_t> position) override;
 	void setMapSize(std::pair<int, int> size) override;
 
 	/**
