+++ display/SDL/arcadeSDL.cpp
@@ -89,7 +89,7 @@ void arcadeSDL::_resizeTexture(SDL_Rect &rect, std::pair<int, int> position)
     rect.h = height / this->_mapSize.second + 0.1;
 }
 
-void arcadeSDL::drawSprite(std::string asset, int color, std::pair<int, int> position) {
+void arcadeSDL::drawSprite(std::string asset, int color, std::pair<size_t, size_t> position) {
     SDL_Surface *surface = IMG_Load(asset.c_str());
     if (!surface) {
         std::cerr << "Error loading image: " << IMG_GetError() << std::endl;
@@ -112,12 +112,12 @@ void arcadeSDL::drawSprite(std::string asset, int color, std::pair<int, int> pos
     SDL_DestroyTexture(texture); // Always destroy the texture after use
 }
 
-void arcadeSDL::drawRectangle(int color, std::pair<int, int> position) {
+void arcadeSDL::drawRectangle(int color, std::pair<size_t, size_t> position) {
     (void)color;
     (void)position;
 }
 
-void arcadeSDL::drawText(std::string text, int color, std::pair<int, int> position) {
+void arcadeSDL::drawText(std::string text, int color, std::pair<size_t, size_t> position) {
     (void)text;
     (void)color;
     (void)position;
