+++ display/SFML/arcadeSFML.hpp
@@ -42,11 +42,11 @@ public:
 
 	void clear() override;
 
-	void drawSprite(std::string asset, int color, std::pair<int, int> position) override;
+	void drawSprite(std::string asset, int color, std::pair<size_t, size_t> position) override;
 
-	void drawRectangle(int color, std::pair<int, int> position) override;
+	void drawRectangle(int color, std::pair<size_t, size_t> position) override;
 
-	void drawText(std::string text, int color, std::pair<int, int> position) override;
+	void drawText(std::string text, int color, std::pair<size_t, size_t> position) override;
 
 	void setMapSize(std::pair<int, int> size) override;
